services/analyzer/requirements.txt
numpy<2
fastapi==0.110.0
uvicorn[standard]==0.27.1
spacy==3.8.2
langchain
langchain-openai
langchain-community
langchain-groq
faiss-cpu==1.8.0
sentence-transformers==2.5.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.3